% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelBeta.R
\name{beta_inc_B}
\alias{beta_inc_B}
\title{Incomplete Beta B(x, a, b)}
\usage{
beta_inc_B(x, a, b)
}
\arguments{
\item{x}{(vector) real number in unit interval}

\item{a}{(scalar) first parameter must be positive real.}

\item{b}{(scalar) second parameter must be real}
}
\value{
B (vector) of incomplete beta evaluated at x
}
\description{
Returns the incomplete beta B(x, a, b) integral.
}
\details{
For parameters (a,b) both positive, this is equal to
pbeta(x,a,b)*beta(a,b). When b nonpositive, we use
the 2F1 function via DLMF 17.07.  We could alternatively
use gsl::beta_inc(a,b,x)*gsl::gsl_sf_beta(a,b), but the
gsl::hyperg_2F1() function is faster and handles b=0.
Unfortunately, for x near 1, the hyperg_2F1 solution
appears to be unstable at b==0.
}
