% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{bispectral_MDtest}
\alias{bispectral_MDtest}
\title{Bispectral MD test}
\usage{
bispectral_MDtest(
  bikernel,
  CVT,
  PIT,
  h_list = NULL,
  informationH = "none",
  singular_Sigma_NA = FALSE
)
}
\arguments{
\item{CVT}{list describing lag structure for each kernel}

\item{PIT}{Vector of probability integral transform values}

\item{h_list}{If provided, then use instead of calling \code{h_closure}.}

\item{informationH}{(string). Specifies information to incorporate in \eqn{H}.}

\item{singular_Sigma_NA}{(boolean). If true, then set \eqn{p}-value to \code{NA} when
\eqn{Sigma} is singular.}

\item{kernlist}{(list) bi-kernel object (see package help)}
}
\value{
\eqn{p}-value of chisq test.
}
\description{
Bispectral MD test
}
