% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralBacktest.R
\name{cross_moment_to_correlation}
\alias{cross_moment_to_correlation}
\title{Cross-moment to correlation.}
\usage{
cross_moment_to_correlation(mux, muA, muB)
}
\arguments{
\item{mux}{(double) cross-moment \eqn{E[\nu_A(P) \nu_B(P)]} under null}

\item{muA}{first two uncentered moments of \eqn{\nu_A(P)}}

\item{muB}{first two uncentered moments of \eqn{\nu_B(P)}}

\item{rho}{linear correlation of \eqn{\nu_A(P)} and \eqn{\nu_B(P)}}
}
\description{
Returns a correlation given a cross-moment and two vectors of uncentered moments \eqn{(\mu_1,\mu_2)} for
each kernel.
}
