% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{cumulants_logisticbeta}
\alias{cumulants_logisticbeta}
\title{Cumulants of the Logistic-beta kernel on unit interval window}
\usage{
cumulants_logisticbeta(param = c(1, 1))
}
\arguments{
\item{param}{beta parameters (a,b)}

\item{support}{lower and upper bound of kernel window}
}
\value{
list with elements \code{cumulants} (list length 4) and scalars \code{skewness} and \code{kurtosis}.
}
\description{
Cumulants of the Logistic-beta kernel on unit interval window
}
