% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{CVT_PtildePower}
\alias{CVT_PtildePower}
\title{Power MD-lag function}
\usage{
CVT_PtildePower(param = 4)
}
\arguments{
\item{param}{power to which \eqn{\abs(2P-1)} is raised}
}
\value{
\eqn{h(P)} MD-lag function
}
\description{
h(P) scaled to have mean 0, variance 1 when P uniform
}
