% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{CVT_twotail}
\alias{CVT_twotail}
\title{Two-sided tail MD-lag function}
\usage{
CVT_twotail(param = 0.99)
}
\arguments{
\item{param}{level for exceedance indicator}
}
\value{
\eqn{h(P)} MD-lag function
}
\description{
Indicator for abs(2 P-1) > abs(2*level-1)
h(P) scaled to have mean 0, variance 1 when P uniform
}
