% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{CVT_uppertail}
\alias{CVT_uppertail}
\title{One-sided upper-tail CVT}
\usage{
CVT_uppertail(param = 0.99)
}
\arguments{
\item{param}{level for exceedance indicator}
}
\value{
\eqn{h(P)} MD-lag function
}
\description{
h(P) scaled to have mean 0, variance 1 when P uniform
}
