% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{.bikernel_correlation}
\alias{.bikernel_correlation}
\title{Generic bikernel correlation function for inheritance of parameter documentation}
\usage{
.bikernel_correlation(support, param)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{list of optional parameters to kernels}
}
\value{
correlation
}
\description{
Generic bikernel correlation function for inheritance of parameter documentation
}
