% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{.monokernel_function}
\alias{.monokernel_function}
\title{Generic singleton kernel function for inheritance of parameter documentation}
\usage{
.monokernel_function(support, param, standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\value{
\eqn{\nu(P)} kernel function
}
\description{
Generic singleton kernel function for inheritance of parameter documentation
}
