% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{.monokernel_moments}
\alias{.monokernel_moments}
\title{Generic singleton kernel moment function for inheritance of parameter documentation}
\usage{
.monokernel_moments(support, param)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\value{
vector of first two uncentered moments
}
\description{
Generic singleton kernel moment function for inheritance of parameter documentation
}
