% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{.tlsfkernel_function}
\alias{.tlsfkernel_function}
\title{Generic TLSF kernel function for inheritance of parameter documentation}
\usage{
.tlsfkernel_function(support, param)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Generic TLSF kernel function for inheritance of parameter documentation
}
