% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{empiricalH}
\alias{empiricalH}
\title{Empirical H function from h_matrix or h_list input}
\usage{
empiricalH(h_matrix)
}
\arguments{
\item{h_matrix}{matrix of lagged \eqn{h} values (including constant column) or list of same}
}
\value{
estimated H matrix
}
\description{
Empirical H function from h_matrix or h_list input
}
