% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{lagmatrix}
\alias{lagmatrix}
\title{Matrix of lagged values of a vector}
\usage{
lagmatrix(x, max.lag)
}
\arguments{
\item{x}{vector}

\item{max.lag}{number of lags to include}
}
\value{
matrix of size (length(x), max.lag+1) of 0-order to max.lag-order lags of x
}
\description{
Matrix of lagged values of a vector
}
