% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{MD_regressor_matrix}
\alias{MD_regressor_matrix}
\title{MD regressor matrix}
\usage{
MD_regressor_matrix(CVT, PIT)
}
\arguments{
\item{CVT}{list describing monokernel or multikernel CVT structure}

\item{PIT}{vector of PIT values (length \eqn{n})}
}
\value{
matrix or list of matrices \eqn{h} (size n x (1+CVT$lags[j]))
}
\description{
Build matrix (or list of matrices) h for the spectral MD test.
}
