% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{monospectral_h_matrix}
\alias{monospectral_h_matrix}
\title{Monospectral h matrix}
\usage{
monospectral_h_matrix(CVT, PIT)
}
\arguments{
\item{CVT}{list describing lag structure}

\item{PIT}{vector of PIT values (length \eqn{n})}
}
\value{
\eqn{h} matrix (size n x (1+monokernel_CVT$lags))
}
\description{
Monospectral h matrix
}
