% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{monospectral_MDtest}
\alias{monospectral_MDtest}
\title{Monospectral MD test}
\usage{
monospectral_MDtest(monokernel, CVT, PIT, h_matrix = NULL,
  informationH = "none", singular_Sigma_NA = FALSE)
}
\arguments{
\item{monokernel}{(list) mono-kernel object (see package help)}

\item{CVT}{list describing lag structure}

\item{PIT}{Vector of probability integral transform values.}

\item{h_matrix}{If provided, then use instead of calling \code{h_closure}.}

\item{informationH}{(string). Specifies information to incorporate in \eqn{H}.}

\item{singular_Sigma_NA}{(boolean). If true, then set \eqn{p}-value to \code{NA} when
\eqn{Sigma} is singular.}
}
\value{
\eqn{p}-value of chisq test.
}
\description{
Monospectral MD test
}
