% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralBacktest.R
\name{monospectral_Ztest}
\alias{monospectral_Ztest}
\title{Spectral Z-test for mono-kernel}
\usage{
monospectral_Ztest(monokernel, PIT, twosided = TRUE)
}
\arguments{
\item{monokernel}{(list) mono-kernel object (see package help)}

\item{PIT}{Vector of probability integral transform values}

\item{twosided}{(boolean) run two-sided test if TRUE, otherwise one-sided}
}
\value{
\eqn{p}-value of Z-test
}
\description{
Spectral Z-test for mono-kernel
}
