% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_arcsin}
\alias{mu_arcsin}
\title{Arcsin moments}
\usage{
mu_arcsin(support = c(0, 1), param = NULL)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
arcsin kernel on the desired \code{support}.Equivalent to mu_beta(param=c(1/2,1/2)).
}
