% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_beta}
\alias{mu_beta}
\title{Beta moments}
\usage{
mu_beta(support = c(0, 1), param = c(1, 1))
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
Beta kernel on the desired \code{support} with parameter \eqn{(p,q)=}\code{param}.
}
