% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_discrete}
\alias{mu_discrete}
\title{Discrete kernel moments}
\usage{
mu_discrete(support = 0.99, param = rep(1, length(support)))
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
discrete kernel on the desired \code{support} with weights in the \code{param} vector.
}
