% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_epanechnikov}
\alias{mu_epanechnikov}
\title{Epanechnikov moments}
\usage{
mu_epanechnikov(support = c(0, 1), param = NULL)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
Epanechnikov kernel on the desired \code{support}.
}
