% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_exponential}
\alias{mu_exponential}
\title{Exponential moments}
\usage{
mu_exponential(support = c(0, 1), param = 0)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
Exponential kernel on the desired \code{support} with parameter \eqn{\zeta=}\code{param}.
}
