% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_linear}
\alias{mu_linear}
\title{Linear moments}
\usage{
mu_linear(support = c(0, 1), param = 1)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
Linear kernel on the desired \code{support}.The slope of the kernel density is given
by \code{param}, which takes values in {-1,1}.
}
