% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_pearson}
\alias{mu_pearson}
\title{Binomial kernel moments for the Pearson test.}
\usage{
mu_pearson(support = NULL, param = 0.99)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
binomial kernel for PIT>\code{param}.
}
\details{
Note that \code{support} is ignored, and \code{param} takes its usual place.
This is to allow the Pearson test to be shoehorned into the multikernel
testing framework.on the desired \code{support} with weights in the \code{param} vector.
}
