% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{mu_probitnormal}
\alias{mu_probitnormal}
\title{Probitnormal score kernel moments}
\usage{
mu_probitnormal(support = c(0, 1), param = 0)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
PNS kernel on the desired \code{support}.  Set \code{param} to control kernel \eqn{j}.
}
