% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralBacktest.R
\name{multispectral_correlation}
\alias{multispectral_correlation}
\title{Theoretical correlation matrix for standardized multikernel}
\usage{
multispectral_correlation(multikernel)
}
\arguments{
\item{multikernel}{(list) multikernel object (see package help)}
}
\value{
\eqn{Sigma} correlation matrix of class Matrix
}
\description{
Theoretical correlation matrix for standardized multikernel
}
