% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{multispectral_h_matrix}
\alias{multispectral_h_matrix}
\title{Multispectral h matrix}
\usage{
multispectral_h_matrix(CVT, PIT)
}
\arguments{
\item{CVT}{list describing bikernel or multikernel lag structure}

\item{PIT}{vector of PIT values (length \eqn{n})}
}
\value{
list of \eqn{h} matrices (size n x (1+CVT$lags[j]))
}
\description{
Multispectral h matrix
}
