% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{multispectral_SUR_MDtest}
\alias{multispectral_SUR_MDtest}
\title{Multispectral SUR MD test}
\usage{
multispectral_SUR_MDtest(
  kernlist,
  CVT,
  PIT,
  h_list = NULL,
  GW = FALSE,
  singular_Sigma_NA = FALSE,
  npval = 0
)
}
\arguments{
\item{kernlist}{(list) multi- or bi-kernel object (see package help)}

\item{CVT}{list describing lag structure for each kernel}

\item{PIT}{Vector of probability integral transform values}

\item{h_list}{If provided, then use instead of calling \code{h_closure}.}

\item{GW}{(boolean) If true, then run Giacomini-White (OLS) in place of SUR.}

\item{singular_Sigma_NA}{(boolean). If true, then set \eqn{p}-value to \code{NA} when
\eqn{Sigma} is singular.}
}
\value{
\eqn{p}-value of chisq test.
}
\description{
Multispectral SUR MD test
}
