% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralBacktest.R
\name{multispectral_Ztest}
\alias{multispectral_Ztest}
\title{Spectral Z-test for multikernel}
\usage{
multispectral_Ztest(multikernel, PIT)
}
\arguments{
\item{multikernel}{(list) multikernel object (see package help)}

\item{PIT}{Vector of probability integral transform values}
}
\value{
\eqn{p}-value of chisq-test
}
\description{
Spectral Z-test for multikernel
}
