% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{nu_arcsin}
\alias{nu_arcsin}
\title{Arcsin kernel function}
\usage{
nu_arcsin(support = c(0, 1), param = NULL, standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\description{
Returns a function nu(P), where nu is the arcsin kernel on
the desired \code{support}. Equivalent to nu_beta(param=c(1/2,1/2)).
}
