% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{nu_discrete}
\alias{nu_discrete}
\title{Discrete kernel function}
\usage{
nu_discrete(support = 0.99, param = rep(1, length(support)),
  standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\description{
Returns a function nu(P), where nu is the discrete kernel on
the desired \code{support} with weights in the \code{param} vector.
}
