% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{nu_gumbel}
\alias{nu_gumbel}
\title{loglog-Gumbel score kernel function}
\usage{
nu_gumbel(support = c(0, 1), param = FALSE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{(boolean) if TRUE, use complementary distribution}
}
\value{
\eqn{\nu(P)} kernel function
}
\description{
Returns a function nu(P), where \eqn{\nu} itself returns a list with the
centered LLS kernels on the desired \code{support}.
}
