% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{nu_linear}
\alias{nu_linear}
\title{Linear kernel function}
\usage{
nu_linear(support = c(0, 1), param = 1, standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\description{
Returns a function nu(P), where nu is the linear kernel on
the desired \code{support}. The slope of the kernel density is given
by \code{param}, which takes values in {-1,1}.
}
