% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{nu_logisticbeta}
\alias{nu_logisticbeta}
\title{Logistic-beta score kernel function}
\usage{
nu_logisticbeta(support = c(0, 1), param = c(1, 1))
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{beta parameters (a,b)}
}
\value{
\eqn{\nu(P)} kernel function
}
\description{
Returns a function nu(P; a,b), where \eqn{\nu} itself returns a list with the
centered LLS kernels on the desired \code{support}.
}
