% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{nu_logitlogistic}
\alias{nu_logitlogistic}
\title{Logit-Logistic score kernel function}
\usage{
nu_logitlogistic(support = c(0, 1), param = NULL)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\value{
\eqn{\nu(P)} kernel function
}
\description{
Returns a function nu(P), where \eqn{\nu} itself returns a list with the
centered LLS kernels on the desired \code{support}.
}
