% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{nu_pearson}
\alias{nu_pearson}
\title{Binomial kernel function for use in Pearson test.}
\usage{
nu_pearson(support = NULL, param = 0.99, standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\description{
Returns a function nu(P), where nu is an indicator for PIT>\code{param}.
Note that \code{support} is ignored, and \code{param} takes its usual place.
This is to allow the Pearson test to be shoehorned into
the multikernel testing framework.
}
