% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{nu_probitnormal}
\alias{nu_probitnormal}
\title{Probitnormal score kernel function}
\usage{
nu_probitnormal(support = c(0, 1), param = 0, standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\description{
Returns a function nu(P), where \eqn{\nu} is the centered PNS j-kernel (\eqn{j=1,2}) on
the desired \code{support}. Set \code{param} to kernel \eqn{j}.
}
