% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{nu_uniform}
\alias{nu_uniform}
\title{Uniform kernel function}
\usage{
nu_uniform(support = c(0, 1), param = NULL, standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\description{
Returns a function nu(P), where nu is the uniform kernel on
the desired \code{support}.
}
