% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{rho_arcsin_epanechnikov}
\alias{rho_arcsin_epanechnikov}
\title{Arcsin/Epanechnikov correlation}
\usage{
rho_arcsin_epanechnikov(support = c(0, 1), param = NULL)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{list of optional parameters to kernels}
}
\description{
Returns the correlation for the Arcsin/Epanechnikov pair on common
\code{support}.
}
