% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{rho_linear_linear}
\alias{rho_linear_linear}
\title{Linear/Linear correlation (positive/negative)}
\usage{
rho_linear_linear(support = c(0, 1), param = NULL)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{list of optional parameters to kernels}
}
\description{
Returns the correlation for the Linear/Linear pair on common
\code{support}.
}
