% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{rho_probitnormal}
\alias{rho_probitnormal}
\title{Probitnormal score correlation}
\usage{
rho_probitnormal(support = c(0, 1), param = NULL)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{list of optional parameters to kernels}
}
\description{
Returns the correlation for VaR exceedance pairs for use in the PNS test.
}
