% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{rho_PtildePower}
\alias{rho_PtildePower}
\title{Correlation of power MD-lag functions}
\usage{
rho_PtildePower(param)
}
\arguments{
\item{param}{list of powers to which \eqn{|2P-1|} is raised}
}
\value{
\eqn{\rho} Correlation of \eqn{h_1(P)} and \eqn{h_2(P)} under null.
}
\description{
Correlation of power MD-lag functions
}
