% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{rHS_arma}
\alias{rHS_arma}
\title{Simulation HS}
\usage{
rHS_arma(n, cormodel)
}
\arguments{
\item{n}{length of sample}

\item{cormodel}{list of \code{ar} and \code{ma} weights, each scalar}
}
\value{
vector of length \code{n}
}
\description{
Simulate a stylized sample for a hypothetical bank using plain-vanilla historical simulation.
Let the unmodelled stochastic volatility be driven by ARMA(1,1) process.
}
