% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{spectral_MDtest}
\alias{spectral_MDtest}
\title{Spectral MD test}
\usage{
spectral_MDtest(kernel, CVT, PIT, h_list = NULL, informationH = "none",
  singular_Sigma_NA = FALSE)
}
\arguments{
\item{kernel}{(list) kernel object (see package help)}

\item{CVT}{(list) CVT object (see package help)}

\item{PIT}{Vector of probability integral transform values}

\item{h_list}{If provided, then use instead of calling \code{h_closure}.}

\item{informationH}{(string). Specifies information to incorporate in \eqn{H}.}

\item{singular_Sigma_NA}{(boolean). If true, then set \eqn{p}-value to \code{NA} when
\eqn{Sigma} is singular.}
}
\value{
\eqn{p}-value of chisq test.
}
\description{
Spectral MD test
}
