% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralBacktest.R
\name{spectral_Ztest}
\alias{spectral_Ztest}
\title{Spectral Z-test for any kernel type}
\usage{
spectral_Ztest(kernel, PIT, twosided = TRUE)
}
\arguments{
\item{kernel}{(list) kernel object (see package help)}

\item{PIT}{Vector of probability integral transform values}

\item{twosided}{(boolean) if monokernel, run two-sided test if TRUE, otherwise one-sided}
}
\value{
\eqn{p}-value of Z-test
}
\description{
Spectral Z-test for any kernel type
}
