% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martingaleDifference.R
\name{theoreticalH}
\alias{theoreticalH}
\title{Theoretical H function from CVT input for bispectral and multispectral MD tests.}
\usage{
theoreticalH(CVT)
}
\arguments{
\item{CVT}{list describing lag structure}
}
\value{
theoretical H matrix
}
\description{
Theoretical H function from CVT input for bispectral and multispectral MD tests.
}
