% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{tlsf_complementary}
\alias{tlsf_complementary}
\title{Complementary distribution functions}
\usage{
tlsf_complementary(
  R,
  Rinverse,
  rho,
  lambdarho,
  B0 = NULL,
  B1 = NULL,
  B2 = NULL
)
}
\arguments{
\item{R}{cdf of distribution}

\item{Rinverse}{inverse cdf of distribution}

\item{rho}{density of distribution}

\item{lambdarho}{\eqn{\lambda_\rho} function}

\item{B0, B1, B2}{integrals expressed as functions of alpha}
}
\value{
named list of complementary functions
}
\description{
Complementary distribution functions
}
