% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{tlsf_scorefunction}
\alias{tlsf_scorefunction}
\title{Score kernel function}
\usage{
tlsf_scorefunction(support = c(0, 1), Rinverse, rho, lambdarho)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{Rinverse}{quantile function of TLSF \eqn{R}}

\item{rho}{density function \eqn{\rho}}

\item{lambdarho}{function \eqn{\lambda_\rho}}
}
\value{
\eqn{\nu(P)} kernel function
}
\description{
Returns a function nu(P), where \eqn{\nu} itself returns a list with the
centered PNS kernels on the desired \code{support}.
}
