% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralBacktest.R
\name{tlsfspectral_Ztest}
\alias{tlsfspectral_Ztest}
\title{Spectral Z-test for tlsfkernel}
\usage{
tlsfspectral_Ztest(tlsfkernel, PIT)
}
\arguments{
\item{tlsfkernel}{(list) TLSF kernel object (see package help)}

\item{PIT}{Vector of probability integral transform values}
}
\value{
\eqn{p}-value of chisq-test
}
\description{
Spectral Z-test for tlsfkernel
}
