% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{vcv_gumbel}
\alias{vcv_gumbel}
\title{loglog-Gumbel variance-covariance function}
\usage{
vcv_gumbel(support = c(0, 1), param = FALSE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{(boolean) if TRUE, use complementary distribution}
}
\value{
\eqn{\Sigma} variance matrix (2x2)
}
\description{
loglog-Gumbel variance-covariance function
}
