% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{vcv_logisticbeta}
\alias{vcv_logisticbeta}
\title{Logistic-beta variance-covariance function}
\usage{
vcv_logisticbeta(support = c(0, 1), param = c(1, 1))
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{beta parameters (a,b)}
}
\value{
\eqn{\Sigma} variance matrix (2x2)
}
\description{
Logistic-beta variance-covariance function
}
