% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTLSF.R
\name{vcv_logitlogistic}
\alias{vcv_logitlogistic}
\title{Logit-Logistic variance-covariance function}
\usage{
vcv_logitlogistic(support = c(0, 1), param = NULL)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}
}
\value{
\eqn{\Sigma} variance matrix (2x2)
}
\description{
Logit-Logistic variance-covariance function
}
