% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelBeta.R
\name{mu_beta}
\alias{mu_beta}
\title{Beta moments}
\usage{
mu_beta(
  support = c(0, 1),
  param = c(1, 1),
  scaling = support,
  force_Mseries = FALSE
)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{scaling}{lower and upper bounds on which to rescale the beta}
}
\description{
Returns a vector of the first two uncentered moments \eqn{(\mu_1,\mu_2)} for the
Beta kernel scaled to \code{scaling} on the desired \code{support}
with parameter \eqn{(p,q)=}\code{param}.  Observe the divergence from
nu_beta() in calling syntax.
}
