% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelBeta.R
\name{nu_beta}
\alias{nu_beta}
\title{Beta kernel function}
\usage{
nu_beta(support = c(0, 1), param, standardize = TRUE)
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{optional parameters to kernel}

\item{standardize}{(boolean) if TRUE, then standardize \eqn{W} to mean 0, variance 1 under the null}
}
\description{
Returns a function nu(P), where nu is the beta kernel on
the desired \code{support} with beta parameters \eqn{(p,q)=}\code{param$beta}
and scaling parameters \code{param$scaling}.  If \code{param} is a vector,
we set \code{scaling} to the \code{support}.
}
