% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelCollection.R
\name{rho_beta_beta}
\alias{rho_beta_beta}
\title{Beta/Beta correlation}
\usage{
rho_beta_beta(support = c(0, 1), param = list(NULL, NULL))
}
\arguments{
\item{support}{lower and upper bound of kernel window}

\item{param}{list of optional parameters to kernels}
}
\description{
Returns the correlation for the Beta/Beta pair on common \code{support}.
}
